+++ crates/sui-adapter/Cargo.toml
@@ -21,12 +21,12 @@ move-core-types.workspace = true
 move-vm-runtime.workspace = true
 move-vm-types.workspace = true
 
-sui-framework = { path = "../sui-framework" }
-sui-verifier = { path = "../sui-verifier" }
-sui-types = { path = "../sui-types" }
-sui-protocol-config = { path = "../sui-protocol-config" }
-workspace-hack = { version = "0.1", path = "../workspace-hack" }
 sui-macros = { path = "../sui-macros" }
+sui-move-natives = { path = "../sui-move-natives" }
+sui-protocol-config = { path = "../sui-protocol-config" }
+sui-types = { path = "../sui-types" }
+sui-verifier = { path = "../sui-verifier" }
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
 
 [dev-dependencies]
 move-package.workspace = true
