+++ crates/sui-config/Cargo.toml
@@ -8,32 +8,32 @@ edition = "2021"
 
 [dependencies]
 anemo.workspace = true
+anyhow = { version = "1.0.64", features = ["backtrace"] }
 bcs = "0.1.4"
 camino = "1.1.1"
-anyhow = { version = "1.0.64", features = ["backtrace"] }
-serde = { version = "1.0.144", features = ["derive", "rc"] }
-serde_with = "2.1.0"
-serde_yaml = "0.8.26"
-rand = "0.8.5"
+csv = "1.2.1"
 dirs = "4.0.0"
-once_cell = "1.16"
-tracing = "0.1.36"
-tempfile = "3.3.0"
-
 fastcrypto.workspace = true
 move-binary-format.workspace = true
 move-core-types.workspace = true
-narwhal-config = { path = "../../narwhal/config" }
+once_cell = "1.16"
+rand = "0.8.5"
+serde = { version = "1.0.144", features = ["derive", "rc"] }
+serde_with = "2.1.0"
+serde_yaml = "0.8.26"
+tempfile = "3.3.0"
+tracing = "0.1.36"
 
-sui-framework = { path = "../sui-framework" }
+narwhal-config = { path = "../../narwhal/config" }
+shared-crypto = { path = "../shared-crypto" }
 sui-adapter = { path = "../sui-adapter" }
-sui-types = { path = "../sui-types" }
+sui-framework = { path = "../sui-framework" }
 sui-keys = { path = "../sui-keys" }
+sui-move-natives = { path = "../sui-move-natives" }
 sui-protocol-config = { path = "../sui-protocol-config" }
 sui-storage = { path = "../sui-storage" }
+sui-types = { path = "../sui-types" }
 workspace-hack = { version = "0.1", path = "../workspace-hack" }
-shared-crypto = { path = "../shared-crypto" }
-csv = "1.2.1"
 [target.'cfg(msim)'.dependencies]
 sui-simulator = { path = "../sui-simulator" }
