+++ crates/sui-core/Cargo.toml
@@ -7,96 +7,96 @@ publish = false
 edition = "2021"
 
 [dependencies]
-rocksdb = "0.20.1"
-dashmap = "5.4.0"
 anyhow = { version = "1.0.64", features = ["backtrace"] }
+arc-swap = "1.5.1"
+async-trait = "0.1.61"
 bcs = "0.1.4"
-chrono = "0.4.23"
-futures = "0.3.23"
 bytes = "1.3.0"
+chrono = "0.4.23"
+dashmap = "5.4.0"
+either = "1.8.0"
 enum_dispatch = "^0.3"
+eyre = "0.6.8"
+futures = "0.3.23"
+itertools = "0.10.4"
+lru = "0.10"
+num_cpus = "1.14.0"
+object_store = "=0.5.4"
+once_cell = "1.16"
+parking_lot = "0.12.1"
+prometheus = "0.13.3"
+rand = "0.8.5"
+rocksdb = "0.20.1"
+scopeguard = "1.1"
 serde = { version = "1.0.144", features = ["derive"] }
 serde_json = "1.0.88"
 serde_with = "2.1.0"
-tokio = { workspace = true, features = ["full", "tracing", "test-util"] }
-tokio-stream = { version = "0.1.8", features = ["sync", "net"] }
-parking_lot = "0.12.1"
-async-trait = "0.1.61"
-tempfile = "3.3.0"
-itertools = "0.10.4"
-tracing = "0.1.36"
 signature = "1.6.0"
-prometheus = "0.13.3"
-arc-swap = "1.5.1"
-object_store = "=0.5.4"
-tokio-retry = "0.3"
-scopeguard = "1.1"
-once_cell = "1.16"
 tap = "1.0"
-either = "1.8.0"
-rand = "0.8.5"
-expect-test = "1.4.0"
-sui-adapter = { path = "../sui-adapter" }
-sui-framework = { path = "../sui-framework" }
-sui-move-build = { path = "../sui-move-build" }
-sui-network = { path = "../sui-network" }
-sui-types = { path = "../sui-types" }
-sui-storage = { path = "../sui-storage" }
-sui-config = { path = "../sui-config" }
-sui-json-rpc-types = { path = "../sui-json-rpc-types" }
-sui-protocol-config = { path = "../sui-protocol-config" }
-lru = "0.10"
+tempfile = "3.3.0"
+thiserror = "1.0.34"
+tokio = { workspace = true, features = ["full", "tracing", "test-util"] }
+tokio-retry = "0.3"
+tokio-stream = { version = "0.1.8", features = ["sync", "net"] }
+tracing = "0.1.36"
 
+fastcrypto.workspace = true
 move-binary-format.workspace = true
 move-bytecode-utils.workspace = true
 move-core-types.workspace = true
-move-vm-runtime.workspace = true
 move-package.workspace = true
-
+move-vm-runtime.workspace = true
 mysten-common.workspace = true
+mysten-network.workspace = true
+telemetry-subscribers.workspace = true
+typed-store-derive.workspace = true
+typed-store.workspace = true
+
 mysten-metrics = { path = "../mysten-metrics" }
 narwhal-config = { path = "../../narwhal/config" }
+narwhal-crypto = { path = "../../narwhal/crypto" }
 narwhal-executor = { path = "../../narwhal/executor" }
 narwhal-network = { path = "../../narwhal/network" }
 narwhal-node = { path = "../../narwhal/node" }
-narwhal-crypto = { path = "../../narwhal/crypto" }
 narwhal-types = { path = "../../narwhal/types" }
 narwhal-worker = { path = "../../narwhal/worker" }
-telemetry-subscribers.workspace = true
-typed-store.workspace = true
-typed-store-derive.workspace = true
-mysten-network.workspace = true
-
-fastcrypto.workspace = true
-workspace-hack = { version = "0.1", path = "../workspace-hack" }
-thiserror = "1.0.34"
-eyre = "0.6.8"
-num_cpus = "1.14.0"
-
-sui-simulator = { path = "../sui-simulator" }
-sui-macros = { path = "../sui-macros" }
 shared-crypto = { path = "../shared-crypto" }
+sui-adapter = { path = "../sui-adapter" }
+sui-config = { path = "../sui-config" }
+sui-framework = { path = "../sui-framework" }
+sui-json-rpc-types = { path = "../sui-json-rpc-types" }
+sui-macros = { path = "../sui-macros" }
+sui-move-build = { path = "../sui-move-build" }
+sui-move-natives = { path = "../sui-move-natives" }
+sui-network = { path = "../sui-network" }
+sui-protocol-config = { path = "../sui-protocol-config" }
+sui-simulator = { path = "../sui-simulator" }
+sui-storage = { path = "../sui-storage" }
+sui-types = { path = "../sui-types" }
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
 
 [dev-dependencies]
-move-symbol-pool.workspace = true
 clap = { version = "3.2.17", features = ["derive"] }
 criterion = { version = "0.4.0" }
+expect-test = "1.4.0"
 fs_extra = "1.3.0"
 more-asserts="0.3.1"
+pretty_assertions = "1.2.1"
 serde-reflection = "0.3.6"
 serde_yaml = "0.8.26"
-pretty_assertions = "1.2.1"
+
+move-symbol-pool.workspace = true
+
 narwhal-test-utils = { path = "../../narwhal/test-utils" }
-sui-verifier = { path = "../sui-verifier" }
 sui-types = { path = "../sui-types", features = ["test-utils"] }
+sui-verifier = { path = "../sui-verifier" }
 
 [target.'cfg(not(target_env = "msvc"))'.dev-dependencies]
 pprof = { version = "0.11.0", features = ["cpp", "frame-pointer"] }
-
-sui-protocol-config = { path = "../sui-protocol-config" }
-sui-macros = { path = "../sui-macros" }
-
 test-fuzz = "3.0.4"
+
+sui-macros = { path = "../sui-macros" }
+sui-protocol-config = { path = "../sui-protocol-config" }
 test-utils = { path = "../test-utils" }
 
 [[example]]
