+++ crates/sui-framework/src/lib.rs
@@ -19,8 +19,6 @@ use sui_types::{
 };
 use tracing::error;
 
-pub mod natives;
-
 /// Represents a system package in the framework, that's built from the source code inside
 /// sui-framework.
 #[derive(Clone, Serialize, PartialEq, Eq, Deserialize)]
