+++ crates/sui-move-natives/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "sui-framework"
+name = "sui-move-natives"
 version = "0.1.0"
 edition = "2021"
 authors = ["Mysten Labs <eng@mystenlabs.com>"]
@@ -8,23 +8,20 @@ license = "Apache-2.0"
 publish = false
 
 [dependencies]
+better_any = "0.1.1"
 bcs = "0.1.4"
-serde = { version = "1.0.144", features = ["derive"] }
-once_cell = "1.16"
-tracing = "0.1.36"
+linked-hash-map = "0.5.6"
+smallvec = "1.9.0"
+
+fastcrypto.workspace = true
+fastcrypto-zkp.workspace = true
 
 sui-types = { path = "../sui-types" }
 
 move-binary-format.workspace = true
 move-core-types.workspace = true
-
+move-stdlib.workspace = true
+move-vm-runtime.workspace = true
+move-vm-types.workspace = true
+sui-protocol-config = { path = "../sui-protocol-config" }
 workspace-hack = { version = "0.1", path = "../workspace-hack" }
-
-[build-dependencies]
-anyhow = { version = "1.0.64", features = ["backtrace"] }
-bcs = "0.1.4"
-
-sui-move-build = { path = "../sui-move-build" }
-
-move-binary-format.workspace = true
-move-package.workspace = true
