+++ crates/sui-move/Cargo.toml
@@ -8,58 +8,55 @@ edition = "2021"
 
 [dependencies]
 anyhow = { version = "1.0.64", features = ["backtrace"] }
-tokio = { workspace = true, features = ["full"] }
-tracing = "0.1.36"
 clap = { version = "3.2.17", features = ["derive"] }
+colored = "2.0.0"
+const-str = "0.5.3"
+git-version = "0.3.5"
+once_cell = { version = "1.16", optional = true }
 serde_json = "1.0.88"
 serde_yaml = "0.8.26"
-git-version = "0.3.5"
-const-str = "0.5.3"
-once_cell = { version = "1.16", optional = true }
-
-sui-core = { path = "../sui-core", optional = true }
-sui-framework = { path = "../sui-framework" }
-sui-move-build = { path = "../sui-move-build" }
-sui-protocol-config = { path = "../sui-protocol-config" }
-sui-types = { path = "../sui-types" }
-sui-cost-tables = { path = "../sui-cost-tables" }
+tracing = "0.1.36"
 
 fastcrypto.workspace = true
-colored = "2.0.0"
-
-telemetry-subscribers.workspace = true
-
 move-binary-format.workspace = true
+move-cli.workspace = true
 move-disassembler.workspace = true
 move-ir-types.workspace = true
-move-prover.workspace = true
+move-package.workspace = true
 move-prover-boogie-backend.workspace = true
+move-prover.workspace = true
 move-unit-test.workspace = true
 move-vm-runtime.workspace = true
 move-vm-test-utils.workspace = true
-move-cli.workspace = true
-move-package.workspace = true
+telemetry-subscribers.workspace = true
+tokio = { workspace = true, features = ["full"] }
+
+sui-core = { path = "../sui-core", optional = true }
+sui-cost-tables = { path = "../sui-cost-tables" }
+sui-move-build = { path = "../sui-move-build" }
+sui-move-natives = { path = "../sui-move-natives" }
+sui-protocol-config = { path = "../sui-protocol-config" }
+sui-types = { path = "../sui-types" }
 
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
 jemalloc-ctl = "^0.5"
 
 [dev-dependencies]
-tempfile = "3.3.0"
-futures = "0.3.23"
-prometheus = "0.13.3"
-
-jsonrpsee = { version = "0.16.2", features = ["jsonrpsee-core"] }
-
-test-utils = { path = "../test-utils" }
-rand = "0.8.5"
-move-package.workspace = true
-sui-core = { path = "../sui-core" }
-sui-node = { path = "../sui-node" }
-sui-macros = { path = "../sui-macros" }
-sui-simulator = { path = "../sui-simulator" }
-mysten-metrics = { path = "../mysten-metrics" }
-
 assert_cmd = "2.0.6"
+futures = "0.3.23"
+jsonrpsee = { version = "0.16.2", features = ["jsonrpsee-core"] }
+prometheus = "0.13.3"
+rand = "0.8.5"
+tempfile = "3.3.0"
+
+move-package.workspace = true
+
+mysten-metrics = { path = "../mysten-metrics" }
+sui-core = { path = "../sui-core" }
+sui-macros = { path = "../sui-macros" }
+sui-node = { path = "../sui-node" }
+sui-simulator = { path = "../sui-simulator" }
+test-utils = { path = "../test-utils" }
 
 [package.metadata.cargo-udeps.ignore]
 normal = ["jemalloc-ctl"]
